backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
boto3==1.34.0
PyJWT==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.0
Werkzeug==3.0.1

